Add failing test cases for region.empty only being called once. Refs #1920
Fix #1920

Add preventDestroy option to region.empty()

Refs #1920 and #1911

Add test from #1918

+++ src/marionette.region.js
@@ -141,10 +141,6 @@ _.extend(Marionette.Region.prototype, Backbone.Events, {
     // We are only changing the view if there is a current view to change to begin with
     var isChangingView = !!this.currentView;
 
-    // Only destroy the current view if we don't want to `preventDestroy` and if
-    // the view given in the first argument is different than `currentView`
-    var _shouldDestroyView = isDifferentView && !preventDestroy;
-
     // Only show the view given in the first argument if it is different than
     // the current view or if we want to re-show the view. Note that if
     // `_shouldDestroyView` is true, then `_shouldShowView` is also necessarily true.
@@ -154,8 +150,8 @@ _.extend(Marionette.Region.prototype, Backbone.Events, {
       this.triggerMethod('before:swapOut', this.currentView);
     }
 
-    if (_shouldDestroyView) {
-      this.empty();
+    if (isDifferentView) {
+      this.empty({ preventDestroy : preventDestroy });
     }
 
     if (_shouldShowView) {
@@ -165,7 +161,7 @@ _.extend(Marionette.Region.prototype, Backbone.Events, {
       // If this happens we need to remove the reference
       // to the currentView since once a view has been destroyed
       // we can not reuse it.
-      view.once('destroy', _.bind(this.empty, this));
+      view.once('destroy', this.empty, this);
       view.render();
 
       if (isChangingView) {
@@ -222,15 +218,24 @@ _.extend(Marionette.Region.prototype, Backbone.Events, {
 
   // Destroy the current view, if there is one. If there is no
   // current view, it does nothing and returns immediately.
-  empty: function() {
+  empty: function(options) {
     var view = this.currentView;
+
+    var emptyOptions     = options || {};
+    var preventDestroy  = !!emptyOptions.preventDestroy;
 
     // If there is no view in the region
     // we should not remove anything
     if (!view) { return; }
 
+    view.off('destroy', this.empty, this);
     this.triggerMethod('before:empty', view);
-    this._destroyView();
+    // Don't destroy the view if we don't want to `preventDestroy`
+    if (!preventDestroy) {
+      this._destroyView();
+    }
+    // Empty the node of HTML
+    this.el.innerHTML='';
     this.triggerMethod('empty', view);
 
     // Remove region pointer to the currentView
